move StartCommercialAsync to Channels class; remove Ads class (#392)
* move StartCommercialAsync to Channels class; remove Ads classs
